refactor: add auth storage contract for session lifecycle
